댓글 API 호출

+++ src/main/resources/static/bootstrap/js/community/Community.js
@@ -1,6 +1,8 @@
 // 커뮤니티 상세보기 JS
 $(function () {
 
+    getReplyData();
+
     let boardId = $("#id").val();
     let actionForm = $("form");
 
@@ -28,9 +30,9 @@ $(function () {
     /* 댓글 전송 */
     $("#review-send-btn").click(function () {
         let text = $("#text").val();
-        let userName = '<%=(String)session.getAttribute("LoginUser")%>';
-        sessionStorage.getItem('LoginUser')
-        console.log(userName);
+
+        let item = sessionStorage.getItem('LoginUser');
+        console.log(item);
         $.ajax({
             url: "/zerogreen/community/" + boardId + "/reply",
             method: "post",
@@ -49,9 +51,9 @@ $(function () {
                 $("#text-count").text("0 / 100");
                 $("#review-table").html("");
                 getReplyData();
+                return false;
             }
         });
-        return false;
     }); // review send end.
 
     // 게시물 삭제
@@ -96,6 +98,7 @@ function textareaResize(obj) {
 
 function getReplyData() {
     let $id = $("#id").val();
+    let $username = $("#login-user").val();
 
     $.getJSON("/zerogreen/community/api/replyList/" + $id, function (list) {
         let html = "";
@@ -111,21 +114,23 @@ function getReplyData() {
             html += "<div class='nest-list'>";
             html += "<input type='hidden' class='replyId' value='" + value.replyId + "'>";
             html += "<span>";
-            html += "<img src='../../static/bootstrap/images/profile/" + value.vegetarianGrade + ".png' width='15px' height='15px'>";
+            html += "<img src='/zerogreen/bootstrap/images/profile/" + value.vegetarianGrade + ".png' width='15px' height='15px'>";
             html += "</span>";
             html += "<span>" + value.nickname + "</span>";
             html += "<span style='font-size: 0.8em; float: right; color: #969696; margin-right: 4px;'>" + value.createdTime + "</span>";
             html += "<div id='text-count-wrapper'>";
-            html += "<span class='cm-text-count' style='display: none;'>0 / 100</span>";
+            html += "<span class='cm-text-count' style='display: none; float:right;'>0 / 100</span>";
             html += "</div>";
-            html += "<p class='re-text'>" + value.text + "</p>";
+            html += "<p class='rp-text' style='margin-bottom: 10px;'>" + value.text + "</p>";
             html += "<a class='modify-test-btn' style='display: none;' onclick='modifyReply(this)'>수정</a>";
-            html += "<div class='reply-btn-wrapper'>";
+            html += "<div class='reply-btn-wrapper' style='margin-bottom: 60px;'>";
             if (value.depth === 1) {
                 html += "<a class='nested-reply-btn' onclick='nestedReplyInput(this)'>답글</a>";
             }
-            html += "<a class='rp-cancel-btn' onclick='deleteReply(" + value.replyId + ")'>삭제</a>";
-            html += "<a class='rp-modify-btn' onclick='replaceTag(this)'>수정</a>";
+            if (value.username === $username) {
+                html += "<a class='rp-cancel-btn' onclick='deleteReply(" + value.replyId + ")'>삭제</a>";
+                html += "<a class='rp-modify-btn' onclick='replaceTag(this)'>수정</a>";
+            }
             html += "</div>";
             html += "<div class='nested-reply-wrapper' style='display: none;'>";
             html += "<textarea class='nested-reply-input' onkeyup='limitTextInput(this)'></textarea>";
@@ -139,7 +144,7 @@ function getReplyData() {
             html += "</div>";
             html += "</div>";
         })
-        $("#review-table").append(html);
+        $("#review-table").html(html);
     });
 } // getReplyData end
 
@@ -168,16 +173,18 @@ function modifyReply(event) {
     let modifyBtn = $(event);
     let replyId = modifyBtn.parent().find(".replyId").val();
     let text = modifyBtn.parent().find(".modify-test");
+    let data = {
+        boardId: boardId,
+        replyId: replyId,
+        text: text.val()
+    }
 
     $.ajax({
-        url: "/zerogreen/replyModify/" + replyId,
-        method: "post",
+        url: "/zerogreen/community/replyModify/" + replyId,
+        method: "put",
         dataType: "json",
-        data: {
-            boardId: boardId,
-            replyId: replyId,
-            text: text.val()
-        }
+        contentType: 'application/json; charset=utf-8',
+        data: JSON.stringify(data)
     })
         .done(function (data) {
             if (data.result === "success") {
@@ -204,16 +211,18 @@ function nestedReplySend(event) {
     let replyId = replyBtn.parents(".comment-wrapper").find(".replyId").val();
     let text = replyBtn.parent().parent().children(".nested-reply-input").val();
 
+    let data = {
+        boardId: boardId,
+        replyId: replyId,
+        text: text
+    }
+    console.log(text);
     $.ajax({
         url: "/zerogreen/community/" + boardId + "/" + replyId + "/nestedReply",
         method: "post",
         dataType: 'json',
         contentType: 'application/json; charset=utf-8',
-        data: {
-            boardId: boardId,
-            replyId: replyId,
-            text: text
-        }
+        data: JSON.stringify(data),
     }).done(function (data) {
         // $("#review-table").replaceWith(fragment);
         if (data.result === "success") {
@@ -225,23 +234,31 @@ function nestedReplySend(event) {
     return false;
 }
 
+/* 댓글 삭제 */
 function deleteReply(replyId) {
 
     let boardId = $("#id").val();
-
-    $.ajax({
-        url: "/zerogreen/community/" + replyId + "/delete",
-        method: "delete",
-        // dataType: "json",
-        data: {
-            replyId: replyId,
-            boardId: boardId
-        },
-    })
-        .done(function (fragment) {
-            alert("댓글이 삭제되었습니다.");
-            $("#review-table").replaceWith(fragment);
-        });
+    let data = {
+        replyId: replyId,
+        boardId: boardId
+    }
+    if (confirm("댓글을 삭제하시겠습니까?") === true) {
+        $.ajax({
+            url: "/zerogreen/community/" + replyId + "/delete",
+            method: "delete",
+            dataType: "json",
+            contentType: "application/json; charset=utf-8",
+            data: JSON.stringify(data),
+        })
+            .done(function (data) {
+                if (data.result === "success") {
+                    $("#review-table").html("");
+                    getReplyData();
+                }
+            });
+    } else {
+        return;
+    }
 }
 
 // 커뮤니티 메인 JS
